Decomp progress: math, oggvorbis, game, system improvements
Major additions to math (Geo, Rot), oggvorbis (mdct, mapping0, synthesis),
and numerous fixes across game code, bandobj, char, rndobj, os, ui, and utl.

+++ src/band3/meta_band/OvershellPanel.cpp
@@ -851,7 +851,7 @@ void OvershellPanel::SetPartRestriction(TrackType t) { mPartRestriction = t; }
 void OvershellPanel::SetMinimumDifficulty(Difficulty d) { mMinimumDifficulty = d; }
 
 void OvershellPanel::ClearTrackTypesFromUsers() {
-    for (int i = 0; i < mSlots.size(); i++) {
+    for (int i = 0; i < (int)mSlots.size(); i++) {
         BandUser *user = mSlots[i]->GetUser();
         if (user && user->IsLocal()) {
             user->SetTrackType(kNumTrackTypes);
